Extract `decodeLines`
