--- src/api/source/mvt/metadata.js
@@ -1,5 +1,0 @@
-export default class Metadata {
-    constructor(columns) {
-        this.columns = columns;
-    }
-}
