+++ src/client/mvt/feature-decoder.js
@@ -8,6 +8,18 @@ export class Polygon {
     }
 }
 
+export function decodeLines(geom, mvt_extent) {
+    let geometry = [];
+    geom.map(l => {
+        let line = [];
+        l.map(point => {
+            line.push(2 * point.x / mvt_extent - 1, 2 * (1 - point.y / mvt_extent) - 1);
+        });
+        geometry.push(line);
+    });
+    return geometry;
+}
+
 /*
     All this clockwise non-sense is needed because the MVT decoder dont decode the MVT fully.
     It doesn't distinguish between internal polygon rings (which defines holes) or external ones, which defines more polygons (mulipolygons)
@@ -136,9 +148,3 @@ function _getPreClippedVertices(geom, mvtExtent) {
         return [x, y];
     });
 }
-
-export default {
-    decodePolygons,
-    isClockWise,
-    clipPolygon
-};
